add test with initial URL
