+++ tests/common/index.ts
@@ -6,7 +6,6 @@ export * from './widgets/pagination';
 export * from './widgets/infinite-hits';
 export * from './widgets/hits';
 export * from './widgets/range-input';
-export * from './widgets/clear-refinements';
 export * from './widgets/hits-per-page';
 export * from './widgets/numeric-menu';
 export * from './widgets/rating-menu';
