--- tests/common/widgets/clear-refinements/routing.ts
@@ -1,57 +1,0 @@
-import { wait } from '@instantsearch/testutils';
-import {
-  createSearchClient,
-  createMultiSearchResponse,
-  createSingleSearchResponse,
-} from '@instantsearch/mocks';
-import type { ClearRefinementsSetup } from '.';
-import { simple } from 'instantsearch.js/es/lib/stateMappings';
-import { history } from 'instantsearch.js/es/lib/routers';
-
-export function createRoutingTests(setup: ClearRefinementsSetup) {
-  describe('routing', () => {
-    describe('URLs created by widget', () => {
-      test('Consistently shows the right URL, even before widget is initialized', async () => {
-        const delay = 100;
-        const router = history();
-        const options = {
-          instantSearchOptions: {
-            indexName: 'indexName',
-            routing: {
-              stateMapping: simple(),
-              router,
-            },
-            searchClient: createSearchClient({
-              search: jest.fn(async (requests) => {
-                await wait(delay);
-                return createMultiSearchResponse(
-                  ...requests.map(() =>
-                    createSingleSearchResponse({
-                      facets: { brand: { Apple: 1 } },
-                    })
-                  )
-                );
-              }),
-            }),
-          },
-          widgetParams: {},
-        };
-
-        await setup(options);
-
-        // Before widgets are completely mounted
-        {
-          // Vue doesn't render anything on first render, so we don't need
-          // to check that the URL is correct.
-          const link = document.querySelector(
-            '[data-testid="ClearRefinements-link"]'
-          );
-          if (link) {
-            // eslint-disable-next-line jest/no-conditional-expect
-            expect(link).toHaveAttribute('href', router.createURL({}));
-          }
-        }
-      });
-    });
-  });
-}
